requirements.txt
langchain>=0.3.10,<0.4.0
langchain-core>=0.3.0
langchain-community>=0.3.10,<0.4.0
langchain-huggingface>=0.1.0
langchain-astradb>=1.0.0
astrapy>=2.0.0
streamlit
python-dotenv
transformers
torch
accelerate
sentence-transformers
pypdf
